+++ contracts/keepers/BOPKeeper.sol
@@ -22,10 +22,10 @@ contract BOPKeeper is IBOPKeeper, BOSSetting, BOPSetting {
     function createPledge(
         // uint32 createDate,
         bytes32 shareNumber,
-        uint256 pledgedPar,
+        uint64 pledgedPar,
         uint40 creditor,
         uint40 debtor,
-        uint256 guaranteedAmt,
+        uint64 guaranteedAmt,
         uint40 caller
     ) external onlyDirectKeeper {
         require(shareNumber.shareholder() == caller, "NOT shareholder");
@@ -45,8 +45,8 @@ contract BOPKeeper is IBOPKeeper, BOSSetting, BOPSetting {
     function updatePledge(
         bytes32 sn,
         uint40 creditor,
-        uint256 pledgedPar,
-        uint256 guaranteedAmt,
+        uint64 pledgedPar,
+        uint64 guaranteedAmt,
         uint40 caller
     ) external onlyDirectKeeper {
         require(pledgedPar > 0, "ZERO pledgedPar");
